refactor: use api for recruitment
